SVM_cuML/requirements.txt
# cuML SVM Requirements
# RAPIDS cuML for GPU-accelerated machine learning
# Python 3.10, 3.11, or 3.12 required

# Core RAPIDS packages (installed via Makefile)
# cuml-cu12==25.12.*
# cudf-cu12==25.12.*

# Additional Python packages
numpy>=1.23.0
cupy-cuda12x>=12.0.0
matplotlib>=3.5.0
seaborn>=0.12.0
scikit-learn>=1.2.0

# Optional: For better visualization
pandas>=1.5.0
